# Conflicts:
#	public/css/app.css
#	public/js/0.js
#	public/js/1.js
#	public/js/10.js
#	public/js/11.js
#	public/js/12.js
#	public/js/13.js
#	public/js/14.js
#	public/js/15.js
#	public/js/16.js
#	public/js/17.js
#	public/js/18.js
#	public/js/19.js
#	public/js/20.js
#	public/js/21.js
#	public/js/22.js
#	public/js/23.js
#	public/js/24.js
#	public/js/25.js
#	public/js/26.js
#	public/js/27.js
#	public/js/28.js
#	public/js/29.js
#	public/js/30.js
#	public/js/31.js
#	public/js/32.js
#	public/js/33.js
#	public/js/34.js
#	public/js/35.js
#	public/js/36.js
#	public/js/37.js
#	public/js/38.js
#	public/js/39.js
#	public/js/4.js
#	public/js/40.js
#	public/js/41.js
#	public/js/42.js
#	public/js/43.js
#	public/js/44.js
#	public/js/45.js
#	public/js/46.js
#	public/js/47.js
#	public/js/48.js
#	public/js/49.js
#	public/js/50.js
#	public/js/51.js
#	public/js/52.js
#	public/js/53.js
#	public/js/54.js
#	public/js/55.js
#	public/js/56.js
#	public/js/57.js
#	public/js/58.js
#	public/js/59.js
#	public/js/60.js
#	public/js/61.js
#	public/js/62.js
#	public/js/63.js
#	public/js/64.js
#	public/js/65.js
#	public/js/66.js
#	public/js/67.js
#	public/js/68.js
#	public/js/8.js
#	public/js/9.js
#	public/js/app.js

+++ app/Http/Controllers/ServerController.php
@@ -2,27 +2,64 @@
 
 namespace App\Http\Controllers;
 
-use App\Jobs\Servers\DeleteServer;
+use App\Mail\Admin\Server\AdminServerCreatedEmail;
 use App\Models\Server;
+use App\Models\User;
 use Illuminate\Http\Request;
+use App\Jobs\Servers\CreateServer;
+use App\Jobs\Servers\DeleteServer;
+use App\Http\Requests\ServerRequest;
+use Illuminate\Support\Facades\Mail;
 use App\Http\Resources\ServerResource;
+use App\Mail\Server\ServerCreatedEmail;
+use App\Http\Requests\ServerUpdateRequest;
 
 class ServerController extends Controller
 {
     public function index()
     {
-        $servers = auth()->user()->servers()->latest()->paginate();
+        $servers = auth()->user()->servers()
+            ->withCount('sites')
+            ->latest()
+            ->paginate();
+
+        $providers = auth()->user()->package ? auth()->user()->package->providers()->pluck('name', 'id') : [];
 
         return inertia('Servers/Index', [
-            'servers' => ServerResource::collection($servers)
+            'servers' => ServerResource::collection($servers),
+            'dataProviders' => $providers
         ]);
     }
 
-    public function store(Request $request)
+    public function store(ServerRequest $request)
     {
-        abort_if(!$request->user()->can('create', Server::class), 403);
+        $provider = $request->user()->package->providers()->findOrFail($request->input('provider'));
+        $region = $provider->regions()->findOrFail($request->input('region'));
+        $plan = $provider->plans()->findOrFail($request->input('plan'));
 
-        // TODO
+        /* @var $server \App\Models\Server */
+        $server = $request->user()->servers()->create([
+            'name' => $request->input('name')
+        ]);
+
+        $server->provider()->associate($provider);
+        $server->providerRegion()->associate($region);
+        $server->providerPlan()->associate($plan);
+        $server->save();
+
+        dispatch(new CreateServer($server));
+
+        Mail::to($request->user())->send(new ServerCreatedEmail($request->user(), $server));
+
+        if (setting('receive_email_on_server_creation')) {
+            $admins = User::query()->where('role', User::ADMIN)->get();
+
+            foreach($admins as $admin){
+                Mail::to($admin)->send(new AdminServerCreatedEmail($request->user(), $server));
+            }
+        }
+
+        return redirect()->route('servers.index');
     }
 
     public function show($id)
@@ -35,9 +72,22 @@ class ServerController extends Controller
         ]);
     }
 
+    public function update(ServerUpdateRequest $request, $id)
+    {
+        $server = $request->user()->servers()->findOrFail($id);
+
+        $server->update([
+            'name' => $request->input('name')
+        ]);
+
+        return redirect()->route('servers.settings.show', $id)->with('success', __('Server information has been updated'));
+    }
+
     public function destroy(Request $request, $id)
     {
         $server = $request->user()->servers()->findOrFail($id);
+
+        $this->authorize('delete', $server);
 
         dispatch(new DeleteServer($server->ploi_id));
 
@@ -49,5 +99,15 @@ class ServerController extends Controller
         $server->delete();
 
         return redirect()->route('servers.index')->with('success', __('Your server is deleted'));
+    }
+
+    public function plansAndRegions(Request $request, $providerId)
+    {
+        $provider = $request->user()->package->providers()->findOrFail($providerId);
+
+        return [
+            'regions' => $provider->regions()->pluck('label', 'id'),
+            'plans' => $provider->plans()->pluck('label', 'id'),
+        ];
     }
 }
